// main/src/domain/models/Instance.d.ts
import type { CommandsFeature, ForwardFeature, MediaFeature, RecallFeature } from '@napgram/feature-kit';
import type { IQQClient } from '../../infrastructure/clients/qq';
import type Telegram from '../../infrastructure/clients/telegram/client';
import ForwardMap from './ForwardMap';
export type WorkMode = 'personal' | 'group' | 'public';
export default class Instance {
    readonly id: number;
    static readonly instances: Instance[];
    private _owner;
    private _isSetup;
    private _workMode;
    private _botSessionId;
    private _qq;
    private _flags;
    private readonly log;
    tgBot: Telegram;
    qqClient?: IQQClient;
    forwardPairs: ForwardMap;
    mediaFeature?: MediaFeature;
    recallFeature?: RecallFeature;
    commandsFeature?: CommandsFeature;
    forwardFeature?: ForwardFeature;
    private featureManager?;
    isInit: boolean;
    private initPromise?;
    eventPublisher?: {
        publishMessageCreated: (...args: any[]) => Promise<void>;
    };
    private constructor();
    private load;
    private init;
    login(botToken?: string): Promise<void>;
    static start(instanceId: number, botToken?: string): Promise<Instance>;
    static createNew(botToken: string): Promise<Instance>;
    get owner(): number;
    get qq(): any;
    get qqUin(): number | undefined;
    get isSetup(): boolean;
    get workMode(): WorkMode;
    get botMe(): import("@mtcute/core").User | undefined;
    get ownerChat(): undefined;
    get botSessionId(): number;
    get flags(): number;
    set owner(owner: number);
    set isSetup(isSetup: boolean);
    set workMode(workMode: WorkMode);
    set botSessionId(sessionId: number);
    set qqBotId(id: number);
    get qqBotId(): number;
    set flags(value: number);
}